refactor: regroupement de compoants dans des dossiers
