+++ src/components/button/button-cercle.tsx
@@ -1,11 +1,11 @@
 import { useState } from 'react'
 import { useNavigate } from 'react-router-dom'
-import Picture from './picture'
+import Picture from '@/components/picture'
 import useLevelStore from '@/store/level-progress'
-import Title from './title'
-import Progressbar from './progress-bar'
-import ProgressNumber from './progress-number'
-import CustomButton from './button-long'
+import Title from '@/components/title'
+import Progressbar from '@/components/progress/progress-bar'
+import ProgressNumber from '@/components/progress/progress-number'
+import CustomButton from '@/components/button/button-long'
 import useExerciseStore from '@/store/data/exercise'
 import { ROUTES } from '@/const'
 import useSound from 'use-sound'
