+++ src/components/card/card-avatar.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from 'react'
 import Picture from '@/components/picture'
 import { useClickAway } from '@uidotdev/usehooks'
-import Title from './title'
-import ChoiceAvatar from './choice-avatar'
-import CustomButton from './button-long'
+import Title from '@/components/title'
+import ChoiceAvatar from '@/components/choice-avatar'
+import CustomButton from '@/components/button/button-long'
 import { API_URL } from '@/config/api'
 import { ROUTES } from '@/const'
 import useAvatarStore from '@/store/user'
