+++ src/components/card/card-correct-answer.tsx
@@ -1,4 +1,4 @@
-import ProgressNumber from '@/components/progress-number'
+import ProgressNumber from '@/components/progress/progress-number'
 
 interface ContentCorrectAnswerProps {
   value: number
