+++ src/components/card/card-valided.tsx
