+++ src/pages/exercise.tsx
@@ -1,9 +1,9 @@
 import Title from '@/components/title'
-import Progressbar from '@/components/progress-bar'
-import ButtonHelp from '@/components/button-help'
-import ProgressNumber from '@/components/progress-number'
-import CardAnswer from '@/components/card-answer'
-import CustomButton from '@/components/button-long'
+import Progressbar from '@/components/progress/progress-bar'
+import ButtonHelp from '@/components/button/button-help'
+import ProgressNumber from '@/components/progress/progress-number'
+import CardAnswer from '@/components/card/card-answer'
+import CustomButton from '@/components/button/button-long'
 import useProgressStore from '@/store/tracking/tracker-progress'
 import useExerciseStore from '@/store/data/exercise'
 import useLevelStore from '@/store/level-progress'
@@ -15,7 +15,7 @@ import useAnswerStore from '@/store/selected-answer'
 import { useNavigate } from 'react-router-dom'
 import useAuthStore from '@/store/tracking/tracker-auth'
 import { ROUTES } from '@/const'
-import CardAudio from '@/components/card-audio'
+import CardAudio from '@/components/card/card-audio'
 import { isAudio } from '@/utils/utils-audio'
 
 export default function Exercice() {
