+++ src/pages/result.tsx
@@ -1,7 +1,7 @@
 import Title from '@/components/title'
-import ContentCorrectAnswer from '@/components/card-correct-answer'
-import CardReward from '@/components/card-reward'
-import CustomButton from '@/components/button-long'
+import ContentCorrectAnswer from '@/components/card/card-correct-answer'
+import CardReward from '@/components/card/card-reward'
+import CustomButton from '@/components/button/button-long'
 import useQuizStore from '@/store/tracking/tracker-answer'
 import useTotalQuizStore from '@/store/tracking/tracker-total-answer'
 import useProgressStore from '@/store/tracking/tracker-progress'
